requirements.txt
openai-whisper
pyannote.audio
torch
torchaudio
librosa
soundfile
numpy
matplotlib
jupyter
ipywidgets
moviepy
ffmpeg-python
